67: * moved hsl generation to helper
